backend/requirements.txt
# ============================================
# requirements.txt
# ============================================
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
websockets==12.0
httpx==0.27.0

# ElevenLabs
elevenlabs>=2.34.0

# Twilio
twilio>=9.0.0

# Google APIs
google-api-python-client>=2.149.0
google-auth-oauthlib>=1.2.1
googlemaps>=4.10.0

# Async & Concurrency
aiofiles>=24.1.0

